+++ packages/matter.js/src/behavior/definitions/level-control/LevelControlServer.ts
@@ -11,7 +11,9 @@ import { Logger } from "../../../log/Logger.js";
 import { StatusCode, StatusResponseError } from "../../../protocol/interaction/StatusCode.js";
 import { TypeFromPartialBitSchema } from "../../../schema/BitmapSchema.js";
 import { Time, Timer } from "../../../time/Time.js";
+import { cropValueRange } from "../../../util/Number.js";
 import { MaybePromise } from "../../../util/Promises.js";
+import { ColorControlServer } from "../color-control/ColorControlServer.js";
 import { GeneralDiagnosticsBehavior } from "../general-diagnostics/GeneralDiagnosticsBehavior.js";
 import { OnOffServer } from "../on-off/OnOffServer.js";
 import { LevelControlBehavior } from "./LevelControlBehavior.js";
@@ -36,12 +38,12 @@ const LevelControlLogicBase = LevelControlBehavior.with(LevelControl.Feature.OnO
  * On-Off Feature is automatically turned on as defined by the matter specification.
  * You should use {@link LevelControlServer.with} to specialize the class for the features your implementation supports.
  *
- * This default implementation also handles the OnOff cluster dependency as defined by the Matter specification
- * automatically.
+ * This default implementation also handles the OnOff cluster dependency and the ColorControl dependency as defined by
+ * the Matter specification automatically.
  *
  * This implementation ignores by default all transition times and sets the level immediately. Alternatively, you can
  * set the `managedTransitionTimeHandling` state attribute to true to have matter.js manage transition times by
- * changing the level value step wise every second. This might be an intermediate solution if you develop
+ * changing the level value step-wise every second. This might be an intermediate solution if you develop
  * independently of defined hardware.
  *
  * If you develop for a specific hardware you should extend the {@link LevelControlServer} class and implement the
@@ -58,15 +60,19 @@ const LevelControlLogicBase = LevelControlBehavior.with(LevelControl.Feature.OnO
  * If you add own implementation you can use:
  * * {@link LevelControlServerLogic.setLevel} to set the level attribute including automatic handling of the onoff dependency
  * * {@link LevelControlServerLogic.setRemainingTime} to set the remaining time attribute when Lighting feature is enabled
+ *
+ * All overridable methods except setRemainingTime can be implemented sync or async by returning a Promise.
  */
 export class LevelControlServerLogic extends LevelControlLogicBase {
     protected declare internal: LevelControlServerLogic.Internal;
     declare state: LevelControlServerLogic.State;
 
+    /** Returns the minimum level, including feature specific fallback value handling. */
     get minLevel() {
         return this.state.minLevel ?? (this.features.lighting ? 1 : 0);
     }
 
+    /** Returns the maximum level, including feature specific fallback value handling. */
     get maxLevel() {
         return this.state.maxLevel ?? 0xfe;
     }
@@ -86,6 +92,14 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
     }
 
     override initialize() {
+        if (this.state.managedTransitionTimeHandling) {
+            this.internal.transitionIntervalTimer = Time.getPeriodicTimer(
+                "LevelControl.step/move",
+                1000,
+                this.callback(this.#stepIntervalTick),
+            );
+        }
+
         if (this.features.lighting) {
             if (this.state.currentLevel === 0) {
                 logger.warn(
@@ -136,13 +150,14 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      * If you want to implement own logic just override {@link moveToLevelLogic} with is also used for {@link moveToLevelWithOnOff}.
      */
     override moveToLevel({ level, transitionTime, optionsMask, optionsOverride }: MoveToLevelRequest) {
-        if (!this.#optionsAllowExecution(optionsMask, optionsOverride)) {
+        const effectiveOptions = this.#calculateEffectiveOptions(optionsMask, optionsOverride);
+        if (!this.#optionsAllowExecution(effectiveOptions)) {
             return;
         }
 
         this.#assertLevelValue(level);
 
-        return this.moveToLevelLogic(level, transitionTime, false);
+        return this.moveToLevelLogic(level, transitionTime, false, effectiveOptions);
     }
 
     /**
@@ -158,26 +173,28 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
     }
 
     /**
-     * Default implementation of the moveToLevel logic. When a transition time is specified the implementation uses the
-     * {@link stepLogic} to manage the move. It also checks if the level is within the min and max level range and sets
+     * Default implementation of the moveToLevel logic. When a transition time is specified the implementation uses a
+     * step based logic to manage the move. It also checks if the level is within the min and max level range and sets
      * the level accordingly. The method uses {@link setLevel} to set the level and handle the on/off state if the method
      * is called by a *WithOnOff command.
      *
      * @param level Level to set
      * @param transitionTime Transition time, ignored in this default implementation
      * @param withOnOff true if the method is called by a *WithOnOff command
+     * @param options Options for the command
      * @protected
      */
-    protected moveToLevelLogic(level: number, transitionTime: number | null, withOnOff: boolean) {
+    protected moveToLevelLogic(
+        level: number,
+        transitionTime: number | null,
+        withOnOff: boolean,
+        options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {},
+    ) {
         // Determine effective transition time
         const transitionTimeValue = transitionTime ?? this.state.onOffTransitionTime ?? null;
 
         // Adjust target level
-        if (this.maxLevel <= level) {
-            level = this.maxLevel;
-        } else if (level <= this.minLevel) {
-            level = this.minLevel;
-        }
+        level = cropValueRange(level, this.minLevel, this.maxLevel);
 
         // If we should move to the new level as fast as possible ...
         if (
@@ -187,17 +204,12 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
             this.currentLevel === level
         ) {
             this.setRemainingTime(0);
-            return this.setLevel(level, withOnOff);
+            return this.setLevel(level, withOnOff, options);
         }
 
         // Else calculate a rate by second and manage the transition
-        const effectiveRate = Math.floor(((level - this.currentLevel) / transitionTimeValue) * 10);
-        return this.#initiateTransition(
-            Math.abs(effectiveRate),
-            effectiveRate < 0 ? LevelControl.StepMode.Down : LevelControl.StepMode.Up,
-            withOnOff,
-            level,
-        );
+        const effectiveRate = Math.ceil(((level - this.currentLevel) / transitionTimeValue) * 10);
+        return this.#initiateTransition(effectiveRate, withOnOff, level, options);
     }
 
     /**
@@ -209,11 +221,12 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      * level is increased or decreased by the step size every second.
      */
     override move({ moveMode, rate, optionsMask, optionsOverride }: MoveRequest) {
-        if (!this.#optionsAllowExecution(optionsMask, optionsOverride)) {
+        const effectiveOptions = this.#calculateEffectiveOptions(optionsMask, optionsOverride);
+        if (!this.#optionsAllowExecution(effectiveOptions)) {
             return;
         }
 
-        return this.moveLogic(moveMode, rate, false);
+        return this.moveLogic(moveMode, rate, false, effectiveOptions);
     }
 
     /**
@@ -237,9 +250,15 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      * @param moveMode Mode (Up/Down) of the move action
      * @param rate Rate of the move action, null if no rate is provided and the default should be used
      * @param withOnOff true if the method is called by a *WithOnOff command
+     * @param options Options for the command
      * @protected
      */
-    protected moveLogic(moveMode: LevelControl.MoveMode, rate: number | null, withOnOff: boolean) {
+    protected moveLogic(
+        moveMode: LevelControl.MoveMode,
+        rate: number | null,
+        withOnOff: boolean,
+        options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {},
+    ) {
         const effectiveRate = rate ?? this.state.defaultMoveRate ?? null;
         if (!this.state.managedTransitionTimeHandling || effectiveRate === null || effectiveRate === 0) {
             // If null rate is requested and also no default rate is set, we should move as fast as possible, so we set
@@ -251,9 +270,14 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
                       ? this.maxLevel
                       : this.minLevel;
             this.setRemainingTime(0);
-            return this.setLevel(level, withOnOff);
+            return this.setLevel(level, withOnOff, options);
         }
-        return this.#initiateTransition(effectiveRate, moveMode as unknown as LevelControl.StepMode, withOnOff);
+        return this.#initiateTransition(
+            effectiveRate * (moveMode === LevelControl.MoveMode.Up ? 1 : -1),
+            withOnOff,
+            undefined,
+            options,
+        );
     }
 
     /**
@@ -261,14 +285,15 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      * After the options checks it uses the {@link stepLogic} method to set the level.
      * If you want to implement own logic just override {@link stepLogic} with is also used for {@link stepWithOnOff}.
      * The logic is implemented as follows: When no transition time is provided, the server will move as fast as
-     * possible, so we set to min/max directly. Else the step logic is applied and the level is increased or decreased
-     * by the step size every transition time interval.
+     * possible, so we set to currentlevel +/- stepSize directly. Else the step logic is applied and the level is
+     * increased or decreased by the step size every transition time interval.
      */
     override step({ stepMode, stepSize, transitionTime, optionsMask, optionsOverride }: StepRequest) {
-        if (!this.#optionsAllowExecution(optionsMask, optionsOverride)) {
+        const effectiveOptions = this.#calculateEffectiveOptions(optionsMask, optionsOverride);
+        if (!this.#optionsAllowExecution(effectiveOptions)) {
             return;
         }
-        return this.stepLogic(stepMode, stepSize, transitionTime, false);
+        return this.stepLogic(stepMode, stepSize, transitionTime, false, effectiveOptions);
     }
 
     /**
@@ -293,6 +318,7 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      * @param stepSize Size of the step action
      * @param transitionTime Time of the step action in 10th of a second
      * @param withOnOff true if the method is called by a *WithOnOff command
+     * @param options Options for the command
      * @protected
      */
     protected stepLogic(
@@ -300,29 +326,33 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
         stepSize: number,
         transitionTime: number | null,
         withOnOff: boolean,
+        options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {},
     ) {
+        const targetLevel = cropValueRange(
+            stepMode === LevelControl.StepMode.Up ? this.currentLevel + stepSize : this.currentLevel - stepSize,
+            this.minLevel,
+            this.maxLevel,
+        );
+
         if (!this.state.managedTransitionTimeHandling || transitionTime === null || transitionTime === 0) {
             // If null/0 transitionTime is requested we should move as fast as possible, so we set to min/max value directly
             this.setRemainingTime(0);
-            return this.setLevel(stepMode === LevelControl.StepMode.Up ? this.maxLevel : this.minLevel, withOnOff);
+            return this.setLevel(targetLevel, withOnOff, options);
         }
-        const effectiveRate = Math.floor((stepSize / transitionTime) * 10);
-        let targetLevel =
-            stepMode === LevelControl.StepMode.Up ? this.currentLevel + stepSize : this.currentLevel - stepSize;
-        if (targetLevel < this.minLevel) {
-            targetLevel = this.minLevel;
-        } else if (targetLevel > this.maxLevel) {
-            targetLevel = this.maxLevel;
-        }
-        return this.#initiateTransition(effectiveRate, stepMode, withOnOff, targetLevel);
+
+        const effectiveRate =
+            Math.ceil((stepSize / transitionTime) * 10) * (stepMode === LevelControl.StepMode.Up ? 1 : -1);
+
+        return this.#initiateTransition(effectiveRate, withOnOff, targetLevel, options);
     }
 
     override stop({ optionsMask, optionsOverride }: StopRequest) {
-        if (!this.#optionsAllowExecution(optionsMask, optionsOverride)) {
+        const effectiveOptions = this.#calculateEffectiveOptions(optionsMask, optionsOverride);
+        if (!this.#optionsAllowExecution(effectiveOptions)) {
             return;
         }
 
-        return this.stopLogic();
+        return this.stopLogic(effectiveOptions);
     }
 
     override stopWithOnOff() {
@@ -335,10 +365,8 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      *
      * @protected
      */
-    protected stopLogic(): MaybePromise<void> {
-        if (this.internal.transitionIntervalTimer?.isRunning) {
-            this.internal.transitionIntervalTimer.stop();
-        }
+    protected stopLogic(_options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {}): MaybePromise<void> {
+        this.internal.transitionIntervalTimer?.stop();
         this.setRemainingTime(0);
     }
 
@@ -362,9 +390,14 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
      *
      * @param level Level which is set by the command
      * @param withOnOff true if the method is called by a *WithOnOff command
+     * @param options Options for the command
      * @protected
      */
-    protected setLevel(level: number, withOnOff: boolean) {
+    protected setLevel(
+        level: number,
+        withOnOff: boolean,
+        options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {},
+    ): MaybePromise<void> {
         const onOffServer =
             this.features.onOff && withOnOff && this.agent.has(OnOffServer) ? this.agent.get(OnOffServer) : undefined;
 
@@ -377,8 +410,16 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
 
         this.state.currentLevel = level;
 
+        let colorSyncResult;
+        // Sync color temperature with level if the feature is enabled and the option is set
+        if (this.features.lighting && options.coupleColorTempToLevel && this.agent.has(ColorControlServer)) {
+            colorSyncResult = this.agent.get(ColorControlServer).syncColorTemperatureWithLevel(level);
+        }
+
         if (onOffServer !== undefined && level > this.minLevel && !onOffServer.state.onOff) {
-            return onOffServer.on();
+            return MaybePromise.then(colorSyncResult, () => onOffServer.on());
+        } else {
+            return colorSyncResult;
         }
     }
 
@@ -389,17 +430,15 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
         const options = this.state.options ?? {};
         return {
             executeIfOff: optionsMask.executeIfOff ? optionsOverride.executeIfOff : options.executeIfOff,
-            coupleColorTempToLevel: optionsMask.coupleColorTempToLevel
-                ? optionsOverride.coupleColorTempToLevel
-                : options.coupleColorTempToLevel,
+            coupleColorTempToLevel: this.features.lighting
+                ? optionsMask.coupleColorTempToLevel
+                    ? optionsOverride.coupleColorTempToLevel
+                    : options.coupleColorTempToLevel
+                : false,
         };
     }
 
-    #optionsAllowExecution(
-        optionsMask: TypeFromPartialBitSchema<typeof LevelControl.Options>,
-        optionsOverride: TypeFromPartialBitSchema<typeof LevelControl.Options>,
-    ) {
-        const options = this.#calculateEffectiveOptions(optionsMask, optionsOverride);
+    #optionsAllowExecution(options: TypeFromPartialBitSchema<typeof LevelControl.Options>) {
         return (
             options.executeIfOff ||
             !this.features.onOff ||
@@ -440,41 +479,41 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
         this.state.currentLevel = this.state.onLevel;
     }
 
-    #initiateTransition(stepSize: number, stepMode: LevelControl.StepMode, withOnOff: boolean, targetLevel?: number) {
-        if (this.internal.transitionIntervalTimer?.isRunning) {
-            this.internal.transitionIntervalTimer.stop();
-        }
+    #initiateTransition(
+        stepSize: number,
+        withOnOff: boolean,
+        targetLevel?: number,
+        options: TypeFromPartialBitSchema<typeof LevelControl.Options> = {},
+    ) {
+        this.internal.transitionIntervalTimer?.stop();
 
         this.internal.currentTransitionData = {
-            changeRate: stepSize * (stepMode === LevelControl.StepMode.Up ? 1 : -1),
+            changeRate: stepSize,
             withOnOff,
             targetLevel,
+            options,
         };
-        logger.debug(`Starting transition interval with stepSize: ${stepSize}, stepMode: ${stepMode}.`);
-        this.internal.transitionIntervalTimer = Time.getPeriodicTimer(
-            "LevelControl.step/move",
-            1000,
-            this.callback(this.#stepIntervalTick),
-        ).start();
+        logger.debug(`Starting transition interval with changeRate: ${this.internal.currentTransitionData.changeRate}`);
+        this.internal.transitionIntervalTimer?.start();
         // Re-Set the current level as start level for the step interval to handle OnOff state changes
-        return this.setLevel(this.currentLevel, withOnOff);
+        return this.setLevel(this.currentLevel, withOnOff, options);
     }
 
     async #stepIntervalTick() {
-        if (!this.internal.currentTransitionData || this.state.currentLevel === null) {
+        if (this.internal.currentTransitionData === undefined || this.state.currentLevel === null) {
             this.internal.transitionIntervalTimer?.stop();
             return;
         }
-        const { changeRate, withOnOff, targetLevel } = this.internal.currentTransitionData;
+        const { changeRate, withOnOff, targetLevel, options } = this.internal.currentTransitionData;
         const newLevel = this.state.currentLevel + changeRate;
         if (newLevel <= this.minLevel) {
             logger.debug(`Stopping transition interval at minLevel: ${this.minLevel}.`);
-            await this.setLevel(this.minLevel, withOnOff);
+            await this.setLevel(this.minLevel, withOnOff, options);
             this.internal.transitionIntervalTimer?.stop();
             this.setRemainingTime(0);
         } else if (newLevel >= this.maxLevel) {
             logger.debug(`Stopping transition interval at maxLevel: ${this.maxLevel}.`);
-            await this.setLevel(this.maxLevel, withOnOff);
+            await this.setLevel(this.maxLevel, withOnOff, options);
             this.internal.transitionIntervalTimer?.stop();
             this.setRemainingTime(0);
         } else {
@@ -482,26 +521,30 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
             if (targetLevel !== undefined) {
                 if (changeRate > 0 && newLevel >= targetLevel) {
                     logger.debug(`Stopping transition interval at targetLevel: ${targetLevel}.`);
-                    await this.setLevel(targetLevel, withOnOff);
+                    await this.setLevel(targetLevel, withOnOff, options);
                     this.internal.transitionIntervalTimer?.stop();
                     this.setRemainingTime(0);
                     return;
                 } else if (changeRate < 0 && newLevel <= targetLevel) {
                     logger.debug(`Stopping transition interval at targetLevel: ${targetLevel}.`);
-                    await this.setLevel(targetLevel, withOnOff);
+                    await this.setLevel(targetLevel, withOnOff, options);
                     this.internal.transitionIntervalTimer?.stop();
                     this.setRemainingTime(0);
                     return;
                 }
             }
             logger.debug(`Setting new level in transition interval: ${newLevel}.`);
-            await this.setLevel(newLevel, withOnOff);
+            await this.setLevel(newLevel, withOnOff, options);
 
             // There is no definition on how often the remaining time should be updated, so we update it with every step
-            if (this.internal.currentTransitionData.changeRate > 0) {
-                this.setRemainingTime(Math.floor(Math.ceil((this.maxLevel - newLevel) / changeRate) * 10));
+            if (changeRate > 0) {
+                this.setRemainingTime(
+                    Math.floor(Math.ceil(((targetLevel ?? this.maxLevel) - newLevel) / changeRate) * 10),
+                );
             } else {
-                this.setRemainingTime(Math.floor(Math.ceil((newLevel - this.minLevel) / changeRate) * 10));
+                this.setRemainingTime(
+                    Math.floor(Math.ceil((newLevel - (targetLevel ?? this.minLevel)) / -changeRate) * 10),
+                );
             }
         }
     }
@@ -514,9 +557,7 @@ export class LevelControlServerLogic extends LevelControlLogicBase {
     }
 
     override async [Symbol.asyncDispose]() {
-        if (this.internal.transitionIntervalTimer?.isRunning) {
-            this.internal.transitionIntervalTimer.stop();
-        }
+        this.internal.transitionIntervalTimer?.stop();
         await super[Symbol.asyncDispose]?.();
     }
 }
@@ -531,6 +572,7 @@ export namespace LevelControlServerLogic {
             changeRate: number;
             withOnOff: boolean;
             targetLevel?: number;
+            options?: TypeFromPartialBitSchema<typeof LevelControl.Options>;
         };
     }